Move coercing pokers to the internal folder
