+++ src/internal/vec.h
@@ -13,5 +13,10 @@ bool r_is_double(sexp* x, r_ssize n, int finite);
 bool r_is_character(sexp* x, r_ssize n);
 bool r_is_raw(sexp* x, r_ssize n);
 
+void r_vec_poke_coerce_n(sexp* x, r_ssize offset,
+                         sexp* y, r_ssize from, r_ssize n);
+void r_vec_poke_coerce_range(sexp* x, r_ssize offset,
+                             sexp* y, r_ssize from, r_ssize to);
+
 
 #endif
