+++ src/rlang/vec.h
@@ -123,11 +123,6 @@ void r_vec_poke_n(sexp* x, r_ssize offset,
 void r_vec_poke_range(sexp* x, r_ssize offset,
                       sexp* y, r_ssize from, r_ssize to);
 
-void r_vec_poke_coerce_n(sexp* x, r_ssize offset,
-                         sexp* y, r_ssize from, r_ssize n);
-void r_vec_poke_coerce_range(sexp* x, r_ssize offset,
-                             sexp* y, r_ssize from, r_ssize to);
-
 extern sexp* r_shared_empty_list;
